OBJ Reader/Writer: avoid precision loss

See merge request OpenMesh/OpenMesh!349

+++ src/OpenMesh/Core/IO/reader/OBJReader.cc
@@ -287,8 +287,8 @@ read_vertices(std::istream& _in, BaseImporter& _bi, Options& _opt,
               std::vector<VertexHandle> & vertexHandles,
               Options & fileOptions)
 {
-    float x, y, z, u, v, w;
-    float r, g, b;
+    double x, y, z, u, v, w;
+    double r, g, b;
 
     std::string line;
     std::string keyWrd;
